requirements.txt
cufflinks==0.17.3
numpy==1.23.2
pandas==1.4.3
plotly==5.10.0
scikit_learn==1.1.2
streamlit==1.12.0
streamlit-aggrid==0.3.3
openpyxl==3.0.10
sklearn == 0.0
matplotlib==3.5.3
matplotlib-inline==0.1.5
XlsxWriter==3.0.3
Pillow==9.2.0
altair<5